# api/tests/postman/inventaio-core-api.openapi.yml
openapi: 3.1.0
info:
  title: InventAI/o API
  description: API del sistema inteligente de consulta y gestión logística para distribuidores de abarrotes.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Docker Compose local
paths:
...
  /api/alertas:
    get:
      tags:
      - Alertas
      summary: Alertas activas por urgencia
      description: 'Genera alertas dinámicas basadas en reglas de inventario. Tipos: inconsistencia_inventario (stock negativo),
        stock_critico, stock_bajo, sin_movimiento, rotacion_baja; la Bodega Central no tiene sin_movimiento ni rotacion_baja
        porque no vende. Urgencia: critica, alta, media. RBAC: admin_sucursal ve solo las de su sucursal. Con page o page_size
        devuelve solo esa página (más page, page_size y pages); sin ellos, todas.'
      operationId: listar_alertas_api_alertas_get
      security:
      - HTTPBearer: []
      parameters:
      - name: tipo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filtrar: inconsistencia_inventario, stock_critico, stock_bajo, sin_movimiento, rotacion_baja'
          title: Tipo
        description: 'Filtrar: inconsistencia_inventario, stock_critico, stock_bajo, sin_movimiento, rotacion_baja'
        example: inconsistencia_inventario
      - name: urgencia
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filtrar: critica, alta, media'
          title: Urgencia
        description: 'Filtrar: critica, alta, media'
        example: alta
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Filtrar por sucursal. admin_sucursal: solo la suya; otra da 403. Inexistente o SIN_SUCURSAL, 422'
          title: Sucursal Id
        description: 'Filtrar por sucursal. admin_sucursal: solo la suya; otra da 403. Inexistente o SIN_SUCURSAL, 422'
        example: 1
      - name: page
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          description: 'Página (INV-26 fix). Sin page ni page_size: todas'
          title: Page
        description: 'Página (INV-26 fix). Sin page ni page_size: todas'
      - name: page_size
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 500
            minimum: 1
          - type: 'null'
          description: Alertas por página; 50 si solo se envía page
          title: Page Size
        description: Alertas por página; 50 si solo se envía page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertaList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/alertas/resumen:
    get:
      tags:
      - Alertas
      summary: Contadores de alertas
      description: Retorna contadores de alertas por urgencia, por sucursal, y por tipo. Con sucursal_id, solo esa ubicación.
      operationId: resumen_alertas_api_alertas_resumen_get
      security:
      - HTTPBearer: []
      parameters:
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Filtrar por sucursal. admin_sucursal: solo la suya; otra da 403. Inexistente o SIN_SUCURSAL, 422'
          title: Sucursal Id
        description: 'Filtrar por sucursal. admin_sucursal: solo la suya; otra da 403. Inexistente o SIN_SUCURSAL, 422'
        example: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertaResumen'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AlertaContadores:
      properties:
        critica:
          type: integer
          title: Critica
        alta:
          type: integer
          title: Alta
        media:
          type: integer
          title: Media
        total:
          type: integer
          title: Total
      type: object
      required:
      - critica
      - alta
      - media
      - total
      title: AlertaContadores
    AlertaItem:
      properties:
        id_producto:
          type: integer
          title: Id Producto
        nombre_producto:
          type: string
          title: Nombre Producto
        categoria:
          type: string
          title: Categoria
        sucursal:
          type: string
          title: Sucursal
        id_sucursal:
          type: integer
          title: Id Sucursal
        tipo_ubicacion:
          type: string
          title: Tipo Ubicacion
        tipo:
          type: string
          title: Tipo
        urgencia:
          type: string
          title: Urgencia
        valor:
          type: number
          title: Valor
        umbral:
          type: number
          title: Umbral
        detalle:
          type: string
          title: Detalle
        fecha:
          type: string
          title: Fecha
      type: object
      required:
      - id_producto
      - nombre_producto
      - categoria
      - sucursal
      - id_sucursal
      - tipo_ubicacion
      - tipo
      - urgencia
      - valor
      - umbral
      - detalle
      - fecha
      title: AlertaItem
    AlertaList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AlertaItem'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page
        page_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page Size
        pages:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pages
        fecha_inventario:
          type: string
          title: Fecha Inventario
      type: object
      required:
      - items
      - total
      - fecha_inventario
      title: AlertaList
    AlertaResumen:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AlertaResumenSucursal'
          type: array
          title: Items
        global_:
          $ref: '#/components/schemas/AlertaContadores'
        por_tipo:
          type: object
          title: Por Tipo
        fecha_inventario:
          type: string
          title: Fecha Inventario
      type: object
      required:
      - items
      - global_
      - por_tipo
      - fecha_inventario
      title: AlertaResumen
    AlertaResumenSucursal:
      properties:
        sucursal:
          type: string
          title: Sucursal
        id_sucursal:
          type: integer
          title: Id Sucursal
        tipo_ubicacion:
          type: string
          title: Tipo Ubicacion
        contadores:
          $ref: '#/components/schemas/AlertaContadores'
      type: object
      required:
      - sucursal
      - id_sucursal
      - tipo_ubicacion
      - contadores
      title: AlertaResumenSucursal
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer